pyproject.toml
[tool.poetry]
name = "nbdefense"
description = "NB Defense CLI."
version = "0.0.0"
authors = ["ProtectAI <community@protectai.com>"]
readme = "README.md"
packages = [{ include = "nbdefense" }]
license = "Proprietary"
classifiers = ["Private :: Do Not Upload"]
exclude = ["tests/*", "Makefile"]

[tool.poetry.scripts]
nbdefense = "nbdefense.cli:cli"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
click = "^8.1.3"
requirements-parser = "0.5.0"
detect-secrets = "1.4.0"
nbformat = "5.6.1"
presidio_analyzer = "2.2.29"
nbconvert = "7.0.0"
jinja2 = "^3.1.2"
requests = "2.28.1"
en_core_web_trf = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.1/en_core_web_trf-3.4.1-py3-none-any.whl" }
tqdm = "4.64.1"
rich = "12.6.0"
pandas = "^1.3.5"
tomlkit = "0.11.6"
torch = [
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp38-cp38-linux_x86_64.whl", python = ">=3.8,<3.9", markers = "sys_platform == 'linux'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp38-cp38-win_amd64.whl", python = ">=3.8,<3.9", markers = "sys_platform == 'win32'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1-cp38-none-macosx_11_0_arm64.whl", python = ">=3.8,<3.9", markers = "sys_platform == 'darwin' and platform_machine == 'arm64'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1-cp38-none-macosx_10_9_x86_64.whl", python = ">=3.8,<3.9", markers = "sys_platform == 'darwin' and platform_machine != 'arm64'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp39-cp39-linux_x86_64.whl", python = ">=3.9,<3.10", markers = "sys_platform == 'linux'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp39-cp39-win_amd64.whl", python = ">=3.9,<3.10", markers = "sys_platform == 'win32'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1-cp39-none-macosx_11_0_arm64.whl", python = ">=3.9,<3.10", markers = "sys_platform == 'darwin' and platform_machine == 'arm64'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1-cp39-none-macosx_10_9_x86_64.whl", python = ">=3.9,<3.10", markers = "sys_platform == 'darwin' and platform_machine != 'arm64'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp310-cp310-linux_x86_64.whl", python = ">=3.10,<3.11", markers = "sys_platform == 'linux'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1%2Bcpu-cp310-cp310-win_amd64.whl", python = ">=3.10,<3.11", markers = "sys_platform == 'win32'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1-cp310-none-macosx_11_0_arm64.whl", python = ">=3.10,<3.11", markers = "sys_platform == 'darwin' and platform_machine == 'arm64'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.12.1-cp310-none-macosx_10_9_x86_64.whl", python = ">=3.10,<3.11", markers = "sys_platform == 'darwin' and platform_machine != 'arm64'" },
]


[tool.poetry.group.test.dependencies]
pytest = "7.1.3"
pytest-watch = "4.2.0"
responses = "0.22.0"
bandit = { version = "1.7.5", extras = ["toml"] }
mypy = "1.1.1"
pandas-stubs = "1.5.3.230304"
types-requests = "2.28.11.15"
types-tqdm = "4.65.0.0"
types-setuptools = "67.6.0.5"

[tool.poetry.group.dev.dependencies]
black = "22.8.0"
pre-commit = "2.20.0"
dunamai = "1.13.2"

[tool.poetry.group.prod.dependencies]
dunamai = "1.13.2"

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
exclude_dirs = ["tests"]
